Make sure the test is ok, and also add that bit of output that was missing to verify that it actually is.
It tests whether hanging 
nodes are identity-constrained to another one; that used to be the case for 
RT elements at one time, but is no longer since they now scale with h; 
consequently, one node is exactly one half of the value of another -- i.e. it 
is still constrained to a single other one, but the weight is not the 
identity any more. The test is therefore ok with its present output.

git-svn-id: https://svn.dealii.org/trunk@13599 0785d39b-7218-0410-832d-ea1e28bc413d