VectorTools::interpolation_to_different_mesh implemented
VectorTools::interpolation_to_different_mesh gives the interpolation of a
dof1-function u1 to a dof2-function u2, where dof1 and dof2 represent
different triangulations with a common coarse grid.

dof1 and dof2 need to have the same finite element discretization.

The mapping between the two spaces is done via the help of the
InterGridMap<> class.

git-svn-id: https://svn.dealii.org/trunk@25013 0785d39b-7218-0410-832d-ea1e28bc413d