Workaround for a problem in LogStream:
It is invalid to test for std::flush or std::endl with function pointers,

  std::ostream & (* const p_flush) (std::ostream &) = &std::flush; p ==
  p_flush ? ...,

as there could be multiple versions of of it.

And in fact, LLVM's libc++ implements std::flush and std::endl in a way
that every shared library and executable has its local copy.

Fix this by implementing a custom streambuf class that queries this sort of
information.

git-svn-id: https://svn.dealii.org/branches/branch_port_to_libcxx@30215 0785d39b-7218-0410-832d-ea1e28bc413d