Fix instantiation issue with FlatManifold properly
This commit reverts the quick fix in commit 982b0683.

This commit closes #56

The issue why we encountered unresolved references to
dealii::FlatManifold<3, 2>::FlatManifold(dealii::Point<2, double>, double)
etc. was due to boilerplate instantiations of Triangulation::get_manifold
for cases with spacedim < dim (in order to make TriaAccessor happy).

This commit fixes the issue by using a nullptr-dereference in order to
return an invalid object - the specific code throws an error message before
that and should actually not be reachable at all.