Silence a warning.
Some newer compilers kept warning about an out-of-bounds array access
in Point<1>::operator() but didn't indicate where exactly it came
from. This turned out to be in DataOutBase::write_eps(), a function
that is implemented only for dim==2 but was compiled also for dim==1
and dim==3 and contained this statement: 'points[vertex](1)'. The patch
reorganizes the code so that the function is only compiled for dim==2
and a more general specialization that does nothing but abort is used
for all other dimensions.