Move declarations in DoFHandler from protected to private.
DoFHandler still had some of its member variables and functions in a protected
section from back when we still had MGDoFHandler derive from it in a meaningful
way. This is now no longer necessary, so remove the protected section and instead
sort things in some sort of natural way into the private section.