Add second template parameter to FE_Nothing.
For historical (and probably not very good) reasons, finite element classes
have a second template argument denoting the space dimension. FE_Nothing never
got this argument and so it wasn't possible to use it in codimension-one
situations. This patch adds the second argument to make it possible to use
it in these situations.

There is not much that one can test this element with, other than ensure that it
compiles. The next commit will actually use the element in a concrete situation.