Remove deprecated functions in ConstraintMatrix.
Specifically, remove the functions that actually create a shrunk linear system, rather than zeroing out rows and columns in an existing linear system. These functions were a bad idea when they were introduced in 1998, and they still are today. We've never used them in anything and they are not used in any of the tutorials or tests.

It is true that I've only deprecated one of the functions yesterday or the day before and remove it today. On the other hand, anyone who would have used that function to shrink a sparsity pattern (which was not previously deprecated) would also have had to use the corresponding function on their matrix (which was already deprecated).