Crank-Nicolson scheme fixed: The linear interpolation between timestep should be upon flux rather than independent vector.
When working at implicit Euler time stepping, the current version
generats an output file that identical to output of the base
version as expected.

The algorithem documentation is also updated while the output
part is not. This is because the correction of energy source
term make the code diverge much sooner and the result doesn't
looks beautiful.