Use AlignedVector::size_type in TableBase::position
Use AlignedVector<T>::size_type as the integer type to carry out the
calculation of the position of the element indicated by the
TableIndices<N> object.

This has the advantage that a potential integer overflow for Tables
with sizes less than max long int in each dimension, but with more than
max long int entries is prevented.