Add an assertion.
Spencer Patty ran into this today: if you try to create a
Trilinos SparsityPattern class where the IndexSets do not
exactly add up, Trilinos somewhere deep down throws an
exception of a type that we do not catch in typical
main() functions (because it is not derived from std::exception).
This leads to impossible to debug problems if you don't happen
to still have the Trilinos sources around and know how to set
breakpoints on all 'throw' statements. Better catch stuff like
this ourselves.