Update documentation in one of the examples.
I tripped over this place the other day because we were setting
boundary indicators without checking that the face is actually
at the boundary. One of my students used the code as the
template to build his own and promptly proceeded to set the
boundary indicator of an internal face.