Reimplement the 2d implementation of edge orientation.
The old implementation, written in 2003 by Mike Anderson, is quite concise
and of high quality. But it is not extensively documented, and is
unnecessarily hard to read because it uses different conventions for names
etc than we use in the rest of the library.

This reimplementation uses standard conventions, and it uses the language and
symbols of a paper about to be submitted that concisely describes the algorithms
and data structures used in this algorithm. As an additional benefit,
it uses the standard vertex ordering of cells rather than the
old-style ordering used in the existing algorithm.

Following the algorithm without understanding the graph theoretical context
of the problem may be difficult, but because it uses the same symbols as in
the paper, should be easy enough if you have the paper.