Add a declared but not implemented function.
This addresses a linker error recently posted to the mailing list. The issue
is that we have iterators that represent faces of cells and a specialization
for the faces of 1d cells, i.e., vertices. These class specializations have
most of the regular accessor members declared, but here is one that was
declared without being defined.

This patch adds a definition of the function. The issue is relatively
uncritical since the so-defined function does not actually have to do
anything other than return an invalid object (vertices have no children).