Remove commented out code in headers. The code in tria_iterator.templates.h was commented out in 2010.
The code in block_sparse_matrix_ez.templates.h was commented out when it
appeared in 2002.

The code in full_matrix.templates.h was commented out in 2001.

The code in precondition_block.templates.h was commented out in 2011.

The code in sparse_matrix_ez.templates.h was commented out in 2003.

The code in mg_block_smoother.h was commented out in 2005.

The code in mg_smoother.h was commented out in 2010.

The code in mg_transfer_component.templates.h was commented out in 2010.

The code in integration_info.templates.h was commented out in 2011.

The code in matrix_tools.h was commented out at some point prior to 2012.

The code in vector_tools.templates.h was commented out at some point prior to 2012.