Testsuite: add configurable executables to DEAL_II_ADD_TEST
This commit adds a second type of tests to the testsuite that consists of a
parameter file "test.prm" with an output file "test.<...>.output". It
assumes that a cmake variable

  TEST_TARGET (or TEST_TARGET_DEBUG and TEST_TARGET_RELEASE)

is defined and will invoke said target with the full path to the parameter
file. (The first type of tests consist of a source file "test.cc" instead,
that is compiled and linked.)