Get rid of FEValuesData in FESystem.
This patch replaces FEValuesData by internal::FEValues::FiniteElementRelatedData
in FESystem. This has two nice benefits:

* We no longer need to copy the mapping related data from the object we get passed
  to the objects that we pass on to the base elements. Rather, we simply reuse the
  one mapping related data object for all base elements.
* This also makes us immune to the situation where a field is added to the mapping
  related data structure. We no longer need to add the copying to FESystem in
  that case.

While there, I took the opportunity to clean up one additional issue: We used to
store an array of pointers to FEValuesData objects. This is not necessary, a
plain array of objects would have sufficed and be more efficient. I made that
change now.