Let the FE classes use FE::InternalDataBase instead of Mapping::InternalDataBase.
For historical reasons, FiniteElement::InternalDataBase is derived from
Mapping::InternalDataBase, but all of the finite element classes need to
use the former. However, we currently pass around pointers to the latter.
There is no good reason for doing so other than 'we can'. Fix this: store
and pass around pointers to FiniteElement::InternalDataBase.