Remove fields from Mapping::InternalDataBase.
It turns out that all of the non-is_first_cell/clear_first_cell() member
variables of Mapping::InternalDataBase are only referenced in Mapping
member functions that are unused throughout the rest of the library.
While they were public, they took an internal object as argument, so
were almost certainly not used anywhere by anyone. Consequently,
zap the fields and the functions that use them.