Remove uses of Mapping::IDB::current_update_flags().
This is the first step towards addressing #1305: MappingCartesian doesn't
actually pre-compute anything at all (update_once is empty), so what
current_update_flags() returns is always equal to update_each. I want to
move to a place where we never use current_update_flags() and in the
fill_fe_*_values() functions only reference update_each. This is simple
for this class, because the two are the same. So replace all occurrences of
current_update_flags() by update_each.