Remove Mapping::update_once/each().
These two functions computed something that, in essence, is only important
for the internal implementation of the mapping classes. It should thus
not be part of the public interface of these classes. Consequently,
remove them and replace them by the only thing that is of interest,
namely to compute transitive closure of the set of flags one needs to
compute -- as now done using the requires_update_flags() function.