Implementation of general cell halo layer function inside GridTools.
Added filtered iterators that work on material id and active FE index, with the
option of only extracting locally owned cells.
Introduced a GridTools function to extract the halo layer that is composed
of a subset of ghost cells (triangulation type dependent).
Multiple tests to check for output based on a general predicate and the
implemented IteratorFilters.
The output of GridTools::compute_ghost_cell_halo_layer is tested against
a distributed triangulation, ensuring that we return all of the ghost cells
on each processor.