Adjust test output.
This test seems to provide different results depending on what version
of NETIS is being used. At least this is suggested by the fact that
the results from the official tester are different than the original
results as added by Denis Davydov. Use the values from the official
'tester' machine.