Bugfix: Remove wrong bugfix
This is just a mistake. We should tackle the fundamental problem, namely
that we do not have templated operator* for mixed std::complex types
available.

Explicitly casting to product_type at every occurence of a possible product
of another scalar with a complex number type is not just tedious, it is
also strictly speaking not correct:

The type trait ProductType<Left, Right>::type defines the type that results
from calling operator* on those types - but neither Left, nor Right, have
to be convertible to a ProductType... (The contraction of tensors as we do
with operator* is an example)