Update test output.
Test outputs change imperceptably and are generally more accurate now
than they were before (see the last column), but there are two or three
points that lie on interfaces between cells and that are now found
for a different cell. This can be seen by noting that for those cells,
the reference coordinates previously had at least one component equal
to zero or one.