Remove a test.
This test previously ran through but returned a point outside the
reference cell. This was due to a fortuitously chosen initial guess
for the Newton iteration (at the center of the cell, pretty stupid)
whereas we now get an exception. In both cases, the function determines
that the point we are seeking lies outside the given cell. As
stated in the documentation, the function in that case either may
return a set of reference coordinates outside the reference cell,
or may throw an exception. The function previously did the former
and now does the latter.

There are three follow-up tests that check the same situation,
so removing this particular test does no harm.